Multi nodes
support multiple nodes with separate log path

+++ src/main/java/com/tz/rsc/controllers/ServiceInfoController.java
@@ -1,33 +1,37 @@
 package com.tz.rsc.controllers;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.tz.rsc.entities.NodeInfo;
 import com.tz.rsc.entities.ServiceInfo;
 import com.tz.rsc.entities.User;
+import com.tz.rsc.repository.NodeInfoRepository;
 import com.tz.rsc.repository.ServiceInfoRepository;
 import com.tz.rsc.repository.UserRepository;
 import com.tz.rsc.utils.Utils;
 
 @RestController
-
-@RequestMapping(value = "/services")
-
+@RequestMapping(value = "/rsc/api/services")
 public class ServiceInfoController
 {
 	private ServiceInfoRepository serviceInfoRepository;
+	private NodeInfoRepository nodeInfoRepository;
 	private UserRepository userRepository;
 
-	public ServiceInfoController(ServiceInfoRepository serviceInfoRepository, UserRepository userRepository)
+	public ServiceInfoController(ServiceInfoRepository serviceInfoRepository, UserRepository userRepository, NodeInfoRepository nodeInfoRepository)
 	{
 		this.serviceInfoRepository = serviceInfoRepository;
 		this.userRepository=userRepository;
+		this.nodeInfoRepository = nodeInfoRepository;
 	}
 	
 	@GetMapping("/all")
@@ -57,10 +61,15 @@ public class ServiceInfoController
 		}
 		
 		if(Utils.isNullOrEmpty(serviceInfo.getName()) ||
-				serviceInfo.getServiceHosts().size()<=0 ||
-				Utils.isNullOrEmpty(serviceInfo.getLogPath())) {
+				serviceInfo.getNodesInfo().size()<=0) {
 			return "{\"message\":\"Missing parameters\"}";
-		}		
+		}
+		
+		for (NodeInfo nodeInfo : serviceInfo.getNodesInfo()) {
+			if(!nodeInfoRepository.findById(nodeInfo.getId()).isPresent()) {
+				return "{\"message\":\"Node does not exists [" + nodeInfo.getId() + "]\"}";
+			}
+		}
 		
 		serviceInfo.setUser(user);
 		serviceInfoRepository.save(serviceInfo);
@@ -70,32 +79,23 @@ public class ServiceInfoController
 	
 	@PostMapping("/update")
 	public String update(@RequestBody ServiceInfo serviceInfo)
-	{		
-		ServiceInfo existingService = this.serviceInfoRepository.findById(serviceInfo.getId()).get();		
+	{
+		User user = this.userRepository.findByUsername(SecurityContextHolder.getContext().getAuthentication()
+                .getPrincipal().toString());
 		
-		if(existingService == null) {
+		Optional<ServiceInfo> existingServiceOp = this.serviceInfoRepository.findById(serviceInfo.getId());		
+		
+		if(!existingServiceOp.isPresent()) {
 			return "{\"message\":\"Service not found\"}";
-		}		
-		
-		if(!Utils.isNullOrEmpty(serviceInfo.getArchiveLogPath())) {
-			existingService.setArchiveLogPath(serviceInfo.getArchiveLogPath());
 		}
 		
-		if(!Utils.isNullOrEmpty(serviceInfo.getLogFileName())) {
-			existingService.setLogFileName(serviceInfo.getLogFileName());
+		ServiceInfo existingService = existingServiceOp.get();
+		
+		if(user == null || existingService.getUser().getId() != user.getId()) {
+			return "{\"message\":\"You are not authorized to update this service\"}";
 		}
 		
-		if(!Utils.isNullOrEmpty(serviceInfo.getLogPath())) {
-			existingService.setLogPath(serviceInfo.getLogPath());
-		}
-		
-		if(!Utils.isNullOrEmpty(serviceInfo.getLogPath())) {
-			existingService.setLogPath(serviceInfo.getLogPath());
-		}
-		
-		if(serviceInfo.getServiceHosts().size() >0) {
-			existingService.setServiceHosts(serviceInfo.getServiceHosts());
-		}
+		existingService.setName(serviceInfo.getName());
 		
 		serviceInfoRepository.save(existingService);
 		return "{\"message\":\"success\"}";
@@ -104,10 +104,82 @@ public class ServiceInfoController
 	@PostMapping("/delete")
 	public String delete(@RequestBody ServiceInfo serviceInfo)
 	{
-		ServiceInfo existingService = this.serviceInfoRepository.findById(serviceInfo.getId()).get();
+		Optional<ServiceInfo> existingServiceOp = this.serviceInfoRepository.findById(serviceInfo.getId());		
+		
+		if(!existingServiceOp.isPresent()) {
+			return "{\"message\":\"Service not found\"}";
+		}
+		
+		ServiceInfo existingService = existingServiceOp.get();
 		
 		
 		serviceInfoRepository.delete(existingService);
+		return "{\"message\":\"success\"}";
+	}
+	
+	@PostMapping("/{id}/nodeinfo/add")
+	public String addNodeInfo(@PathVariable("id") long serviceId, @RequestBody NodeInfo nodeInfo)
+	{
+		User user = this.userRepository.findByUsername(SecurityContextHolder.getContext().getAuthentication()
+                .getPrincipal().toString());
+		
+		Optional<ServiceInfo> existingServiceOp = this.serviceInfoRepository.findById(serviceId);		
+		
+		if(!existingServiceOp.isPresent()) {
+			return "{\"message\":\"Service not found\"}";
+		}	
+		
+		ServiceInfo existingService = existingServiceOp.get();
+		
+		if(user == null || existingService.getUser().getId() != user.getId()) {
+			return "{\"message\":\"You are not authorized to update this service\"}";
+		}
+		
+		Optional<NodeInfo> existingNodeOp = this.nodeInfoRepository.findById(nodeInfo.getId());		
+		
+		if(!existingNodeOp.isPresent()) {
+			return "{\"message\":\"Node not found\"}";
+		}	
+		
+		NodeInfo existingNode = existingNodeOp.get();
+		
+		if(user == null || existingNode.getUser().getId() != user.getId()) {
+			return "{\"message\":\"You are not authorized to use this node\"}";
+		}
+		
+		existingService.getNodesInfo().add(existingNode);
+		
+		serviceInfoRepository.save(existingService);
+		return "{\"message\":\"success\"}";
+	}	
+	
+	@PostMapping("/{id}/nodeinfo/delete")
+	public String deleteNodeInfo(@PathVariable("id") long serviceId, @RequestBody NodeInfo nodeInfo)
+	{
+		User user = this.userRepository.findByUsername(SecurityContextHolder.getContext().getAuthentication()
+                .getPrincipal().toString());
+		
+		Optional<ServiceInfo> existingServiceOp = this.serviceInfoRepository.findById(serviceId);		
+		
+		if(!existingServiceOp.isPresent()) {
+			return "{\"message\":\"Service not found\"}";
+		}	
+		
+		ServiceInfo existingService = existingServiceOp.get();
+		
+		if(user == null || existingService.getUser().getId() != user.getId()) {
+			return "{\"message\":\"You are not authorized to update this service\"}";
+		}		
+		
+		NodeInfo existingNode = existingService.getNodeInfoById(nodeInfo.getId());
+		
+		if(existingNode == null) {
+			return "{\"message\":\"Node not found\"}";
+		}	
+		
+		existingService.getNodesInfo().remove(existingNode);
+		
+		serviceInfoRepository.save(existingService);
 		return "{\"message\":\"success\"}";
 	}
